project/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
transformers==4.37.0
torch==2.1.2
torchaudio==2.1.2
spotipy==2.23.0
python-dotenv==1.0.0
librosa==0.10.1
soundfile==0.12.1
